streaming/producer/requirements.txt
kafka-python-ng
yfinance
pandas
